refactor(ui5-select): migrate to TypeScript - IOption interface moved from Option.ts to Select.ts
